update product module to camelcase namespacing
