+++ client/modules/product/actions/productSingleActions.js
