+++ client/modules/product/components/CreateProduct.js.jsx
@@ -3,13 +3,13 @@ import Base from "../../../global/components/BaseComponent.js.jsx";
 import { connect } from 'react-redux';
 
 // import actions
-import * as singleActions from '../actions/single';
+import * as singleActions from '../actions/productSingleActions';
 
 
 // import components
 import ProductForm from './ProductForm.js.jsx';
 
-class Create extends Base {
+class CreateProduct extends Base {
   constructor(props) {
     super(props);
     this.state = this.props;
@@ -74,7 +74,7 @@ class Create extends Base {
   }
 }
 
-Create.propTypes = {
+CreateProduct.propTypes = {
   dispatch: PropTypes.func.isRequired
 }
 
@@ -89,4 +89,4 @@ const mapStateToProps = (state) => {
 
 export default connect(
   mapStateToProps
-)(Create);
+)(CreateProduct);
