+++ client/modules/product/productRoutes.js.jsx
@@ -4,11 +4,11 @@ import React from 'react';
 import { Route, IndexRoute } from 'react-router';
 
 //import Components
-import Layout from './components/Layout.js.jsx';
-import List from './components/List.js.jsx';
-import Single from './components/Single.js.jsx';
-import Create from './components/Create.js.jsx';
-import Update from './components/Update.js.jsx';
+import Layout from './components/ProductLayout.js.jsx';
+import List from './components/ProductList.js.jsx';
+import Single from './components/SingleProduct.js.jsx';
+import Create from './components/CreateProduct.js.jsx';
+import Update from './components/UpdateProduct.js.jsx';
 
 const productRoutes =
 <Route path="/products" component={Layout} >
